run-command API: rename "env_array" to "env"
As noted in the preceding commit it would be nice to not deal with the
oddity of having an "_array" suffix for the "env" strvec, but not for
its "args" sibling.

We can do that with coccinelle in a way that's easy to review. The
only manual changes in this commit are:

 * git mv contrib/coccinelle/run_command{.pending,}.cocci
 * The change in run-command.h to rename the struct member itself, and
   to change its initializer.

The rest of this is all the result of a subsequent:

 * make contrib/coccinelle/run_command.cocci.patch
 * patch -p1 <contrib/coccinelle/run_command.cocci.patch
 * git add -u

I've avoided changing any comments and derived variable names here,
that will all be done in the next commit.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>