cocci: remove env_array -> env migration
Remove the coccinelle rule renaming "env_array" to "env". In the
preceding commit we carried out this rename based on this rule, anyone
with out-of-tree patches still referring to "env" will now get a
compilation error.

We could keep the rule to help those developers change their code by
running "make coccicheck", but as the solution is obvious in this case
let's not leave this stale rule in place.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>